Reworked block allocator.
Now requires memory needed to be specified at sovler creation.
Previously, the library allocated an humongous amount of memory and required "finalize()" to be called, which reallocated to the actually required size.
I assume the idea was that this is relatively quick to run, easy to implement and shrinking memory shouldn't move the memory relocation.

However, I encountered that on MacOS memory was always moved and thus the code would not run. In any case, reallocate() does not give any guarantee for the code to remain at the same memory location.
Simply throwing an exception and terminating is not really solution, though it seems to work on Linux/Windows for what it's worth.

Requiring pre-calculating the memory is a bit of an overhead and pretty anoying, but should be a more stable solution.

Sadly, there was no good way to write the code for estimation into the current codebase without a major rewrite of the interface.
At least the decomposition would have had to be done internally as well, however, one can make an argument for libmpopt to represent the pure graphical model reformulation of the original QAP problem.

I therefore decided to leave everything as is as and added the adapter to the parent project. Though it is very unintuitive to require the caller to estimate the memory needed for the solver.