Removed finalize() function from allocator class.
Finalize was used to resize arena at the end of allocation stage. This does not work as expected cross-platform and would need a major redesign of memory management. At the moment, following the previous commit, it is accepted that the virtual memory stays reserved during execution. Furthermore, it was reduced to block at most 256GB of virtual memory. Though there may be edge cases with many processes spawned in parallel, where this implementation will fail, it should be good enough for now.